Rename the android support library to native_support to make it clear that shared support code that does not depend on JNA should go elsewhere
