+++ components/places/android/src/main/java/mozilla/appservices/places/LibPlacesFFI.kt
@@ -12,8 +12,8 @@ import com.sun.jna.Pointer
 import com.sun.jna.PointerType
 import com.sun.jna.StringArray
 import java.lang.reflect.Proxy
-import mozilla.appservices.support.RustBuffer
-import mozilla.appservices.support.loadIndirect
+import mozilla.appservices.support.native.RustBuffer
+import mozilla.appservices.support.native.loadIndirect
 import org.mozilla.appservices.places.BuildConfig
 
 @Suppress("FunctionNaming", "FunctionParameterNaming", "LongParameterList", "TooGenericExceptionThrown")
