+++ components/push/android/src/main/java/mozilla/appservices/push/LibPushFFI.kt
@@ -9,9 +9,9 @@ import android.util.Log
 import com.sun.jna.Library
 import com.sun.jna.Native
 import com.sun.jna.Pointer
-import mozilla.appservices.support.RustBuffer
+import mozilla.appservices.support.native.RustBuffer
 import java.lang.reflect.Proxy
-import mozilla.appservices.support.loadIndirect
+import mozilla.appservices.support.native.loadIndirect
 import org.mozilla.appservices.push.BuildConfig
 
 @Suppress("FunctionNaming", "FunctionParameterNaming", "LongParameterList", "TooGenericExceptionThrown")
