+++ components/support/android/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.mozilla.appservices.support" />
+    package="org.mozilla.appservices.native_support" />
