+++ components/viaduct/android/src/main/java/mozilla/appservices/httpconfig/LibViaduct.kt
@@ -9,8 +9,8 @@ import com.sun.jna.Library
 import com.sun.jna.Callback
 import com.sun.jna.Native
 import java.lang.reflect.Proxy
-import mozilla.appservices.support.RustBuffer
-import mozilla.appservices.support.loadIndirect
+import mozilla.appservices.support.native.RustBuffer
+import mozilla.appservices.support.native.loadIndirect
 import org.mozilla.appservices.httpconfig.BuildConfig
 
 @Suppress("FunctionNaming", "TooGenericExceptionThrown")
